+++ config/routes.rb
@@ -38,18 +38,18 @@ Rails.application.routes.draw do
     end
   end
 
-  resources :reviews do
+  resources :challenges do
     member do
-      post :create_comment
-      post :preview
-      post :show_comment
-      get :edit_comment
-      post :update_comment
-      post :destroy_comment
-      post :vote
-      post :update_tags
+      get "images/*path", action: 'images', as: :images, format: false
+      get "signal/:signal", action: 'signal', as: :signal
     end
   end
+
+  scope module: 'challenges' do
+    get "challenges/flow-limitation/review", action: 'review', as: :flow_limitation_challenge_review
+    get "challenges/flow-limitation/submitted", action: 'submitted', as: :flow_limitation_challenge_submitted
+  end
+
 
   resources :datasets do
     member do
@@ -91,6 +91,19 @@ Rails.application.routes.draw do
     end
   end
 
+  resources :reviews do
+    member do
+      post :create_comment
+      post :preview
+      post :show_comment
+      get :edit_comment
+      post :update_comment
+      post :destroy_comment
+      post :vote
+      post :update_tags
+    end
+  end
+
   scope module: 'static' do
     get :demo
     get :parallax
@@ -100,16 +113,6 @@ Rails.application.routes.draw do
     get "showcase(/:slug)", action: 'showcase', as: :showcase
     get :version
     get :sizes
-  end
-
-  get 'challenges' => "challenges#flow_limitation"
-
-  scope module: 'challenges' do
-    get "challenges/flow-limitation", action: 'flow_limitation', as: :flow_limitation_challenge
-    get "challenges/flow-limitation/step/1", action: 'step1', as: :flow_limitation_challenge_step1
-    get "challenges/flow-limitation/step/2", action: 'step2', as: :flow_limitation_challenge_step2
-    get "challenges/flow-limitation/review", action: 'review', as: :flow_limitation_challenge_review
-    get "challenges/flow-limitation/submitted", action: 'submitted', as: :flow_limitation_challenge_submitted
   end
 
   resources :tools do